Moved profile icon to Avatar component
