+++ frontend/src/routes/(app)/+layout.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
     import { page } from "$app/state";
     import type { LayoutData } from "./$types";
-    import sampleAvatar from "$lib/assets/images/sample_avatar.jpg";
 
     import type { Snippet } from "svelte";
     import Logo from "$lib/components/Logo.svelte";
+    import Avatar from "$lib/components/Avatar.svelte";
 
     interface Props {
         children?: Snippet;
@@ -18,28 +18,12 @@
     let popover: HTMLDivElement;
     let popoverVisible = $state(false);
 
-    function initials(name: string) {
-        const words = name.split(" ").filter((part) => part.length > 0);
-        const initialsArray = words.map((part) => part.charAt(0).toUpperCase());
-        return initialsArray.join("");
-    }
-
     function handleClick(e: MouseEvent) {
         if (popover && !popover.contains(e.target as Node)) {
             popoverVisible = false;
         }
     }
 </script>
-
-{#snippet profileIcon(size: string)}
-    <div class="icon" style:--size={size}>
-        {#if profile.username || profile.full_name}
-            <span>{initials(profile.full_name ?? profile.username)}</span>
-        {:else}
-            <img src={sampleAvatar} alt="Generic user icon" />
-        {/if}
-    </div>
-{/snippet}
 
 <div id="container">
     <nav>
@@ -51,11 +35,11 @@
                 popoverVisible = !popoverVisible;
             }}
         >
-            {@render profileIcon("50px")}
+            <Avatar {profile} size={50} />
         </button>
         <div bind:this={popover} id="popover" class:visible={popoverVisible}>
             <section id="details">
-                {@render profileIcon("50px")}
+                <Avatar {profile} size={50} />
                 <div>
                     <p><strong>{profile.full_name}</strong></p>
                     <p>{profile.username}</p>
@@ -66,7 +50,11 @@
                 <li><a href="/studio">Studio</a></li>
                 <li><a href="/listening">Listening Data</a></li>
                 <li><a href="/profile">Profile</a></li>
-                <li><a href="/logout">Log Out</a></li>
+                <li>
+                    <a href="/logout" data-sveltekit-preload-data="off">
+                        Log Out
+                    </a>
+                </li>
             </ul>
         </div>
     </nav>
@@ -148,30 +136,8 @@
         }
     }
 
-    .icon {
-        border-radius: 50%;
-        color: var(--fg);
-        background-color: var(--bg-subtle);
-        border: 1px solid var(--border);
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        width: var(--size);
-        height: var(--size);
-
-        img {
-            width: 100%;
-            height: 100%;
-            object-fit: cover;
-        }
-
-        span {
-            font-size: calc(var(--size) * 0.4);
-        }
-    }
-
     .content {
-        margin: 20px;
+        margin: 1em;
         flex-grow: 1;
         min-height: 0;
     }
